--- backend/src/core/agent/graph_annotations.py
@@ -1,7 +1,0 @@
-class ResetOnStart:
-    """
-    Annotation to mark fields in a GraphState that should be reset to their default value at the
-    start of a graph run.
-    """
-
-    pass
