+++ backend/tests/fixtures/doc.py
@@ -191,7 +191,6 @@ async def ingested_doc_table(doc_table, readonly_doc_set_table, api_server, sql_
 
         yield doc_table
 
-
         with sql_sessionmaker() as session:
             stmt = select(doc_table).where(doc_table.id.in_(doc_ids))
             result = session.execute(stmt).fetchall()
@@ -206,8 +205,8 @@ async def ingested_doc_table(doc_table, readonly_doc_set_table, api_server, sql_
         logger.info('doc statuses after test: %s', [str(x) for x in statuses])
 
         import os
-        os._exit(1)
 
+        os._exit(1)
 
     finally:
         # Clean up table after we are done.
